removed duplicate docs for scripts & added help options to all scripts

+++ x
@@ -6,5 +6,16 @@
 # x - Exit shortcut
 #
 # Usage: x
+#
+# Behavior:
+#	Ends the process running the command. Useful for exiting a terminal 
+#	with a single character command.
 #?
+
+# Help option
+if [[ "$1" =~ ^(--help|-h)$ ]]; then
+	show-help "$0"
+	exit 1
+fi
+
 kill -9 "$PPID"
